Add namespace to engines tests
